feat: custom formatting via treefmt
Having code formatting configured standalone on the command line feels
more flexible than in-editor. I've went with treefmt, and specifically
treefmt-nix, which allows managing dependencies (such as the formatters)
using Nix.

Emacs Lisp is trickier as there doesn't appear to be a good standalone
formatter available, with most of the logic baked into Emacs itself. For
that reason I've packaged lisp-format, a Bash and Elisp script intended
for execution via Emacs, as a NUR package and configured it here as a
custom formatter.

Refs: #12