feat: Google Calendar integration
Using org-gcal to integrate with Google Calendar, meaning my schedule is
visible in org alongside all of my other todos.

This also demonstrates how to handle secrets (in this case, the OAuth
client secret needed to call the Google Calendar APIs and the OAuth
access and refresh tokens we receive in response). This is implemented
using Emacs Auth Source library, which has good support for GPG for the
encryption.

Refs: #7, #16