fix: forge not checking all auth-sources
I'm unsure why this is happening, but Magit Forge appears to be
sensitive to the order in which `auth-sources` is defined. In the
previous commit, the GPG key used to store the OAuth client secret for
org-gcal was first in the `auth-sources` list and this led to Forge
claiming it couldn't find the GitHub access key even though it's in
the macOS keychain which is also defined in `auth-sources`.

Refs: #16