feat: linting commit messages
I've integrated commitlint to lint commit messages against the
Conventional Commits standard. I've opted to store the machinery for
this in an internal/ directory instead of having it all in the
repository root.

Currently I have this enabled locally using the commit-msg git hook
included in this commit, but I'll enable this in GitHub Actions too.

Refs: #15