feat: improve linting configuration
I've removed the custom Lisp formatter as I wasn't happy with the
output. I have a development branch with some further work on that, but
given how little Lisp code I have currently it's not a priority.

I also refactored the actual Treefmt configuration out of the main
flake.nix and into a standalone file in the internal/ directory.

Refs: #21